+++ lib/src/values/enumeration.dart
@@ -27,7 +27,22 @@ enum MessageType {
 
   /// Only supported on android and ios
   voice,
-  custom
+  custom;
+
+  static MessageType? tryParse(String? value) {
+    final type = value?.trim().toLowerCase();
+    if (type?.isEmpty ?? true) return null;
+    if (type == image.name) {
+      return image;
+    } else if (type == text.name) {
+      return text;
+    } else if (type == voice.name) {
+      return voice;
+    } else if (type == custom.name) {
+      return custom;
+    }
+    return null;
+  }
 }
 
 /// Events, Wheter the user is still typing a message or has
@@ -36,7 +51,27 @@ enum TypeWriterStatus { typing, typed }
 
 /// [MessageStatus] defines the current state of the message
 /// if you are sender sending a message then, the
-enum MessageStatus { read, delivered, undelivered, pending }
+enum MessageStatus {
+  read,
+  delivered,
+  undelivered,
+  pending;
+
+  static MessageStatus? tryParse(String? value) {
+    final status = value?.trim().toLowerCase();
+    if (status?.isEmpty ?? true) return null;
+    if (status == read.name) {
+      return read;
+    } else if (status == delivered.name) {
+      return delivered;
+    } else if (status == undelivered.name) {
+      return undelivered;
+    } else if (status == pending.name) {
+      return pending;
+    }
+    return null;
+  }
+}
 
 /// Types of states
 enum ChatViewState { hasMessages, noData, loading, error }